change harvest block field to long2
(cherry picked from commit da6449d785284d67a57f6d30440a70f0f0808fd3)

+++ src/main/java/pro/belbix/ethparser/model/HarvestTx.java
@@ -42,7 +42,7 @@ public class HarvestTx implements EthTransactionI {
         HarvestDTO dto = new HarvestDTO();
         dto.setId(hash + "_" + logId);
         dto.setHash(hash);
-        dto.setBlock(block);
+        dto.setBlock(block.longValue());
         dto.setVault(removeBracers(Vaults.vaultHashToName.get(vault.getValue())));
         dto.setConfirmed(success);
         dto.setMethodName(methodName);
